% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forint.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{Get BTC price}
\usage{
get_bitcoin_price(retried = 0)
}
\arguments{
\item{retried}{number of times previously retried (and failed) the API query}
}
\value{
number
}
\description{
Get BTC price
}
